pyproject.toml
[project]
name = "exchange-latencies-profiler"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "eventkit",
    "fastapi[standard]>=0.116.1",
    "psycopg>=3.2.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.42",
    "sqlmodel>=0.0.24",
    "tenacity>=9.1.2",
    "websocket-client>=1.8.0",
]

[tool.uv.sources]
eventkit = { git = "https://github.com/erdewit/eventkit.git" }
